Divided test_models.py into two separate scripts.
